requirements.txt
click==8.1.3
distlib==0.3.4
filelock==3.7.0
Flask==2.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
platformdirs==2.5.2
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-110/six-1.15.0-py2.py3-none-any.whl
virtualenv==20.14.1
Werkzeug==2.1.2
zipp==3.8.0

pip==22.1.1
SNAKES~=0.9.29
python-dateutil~=2.8.2